requirements.txt
streamlit == 1.14.0
DLite-Python == 0.3.15
numpy == 1.23.2
pandas == 1.4.3